A silent stop helper function

This function mutes the Error when calling stop Palette creating function

palette_create(num.color = 5, df.rgb, threshold = 25, plot = T, proceed = NA)

Arguments

num.color

An integer. How many colors needed in the palette

df.rgb

A data.frame. This dataframe should have all pixels' rgb information of an image

threshold

A number. To determine the minimum accepted color difference in the palette

plot

T/F. Default to display the colors.

proceed

A testing parameter

Value

A dataframe. Hex codes and their rgb values.

Details

This function creates a palette from an image. The image need to be input in dataframe of pixels in rgb. If the extracted colors are sufficient for differentiating it automatically generates the pallete, but if the color contrast is less than differentiating threshold, it will throw warning before proceeding.

Examples

df.rgb = image2rgb()
palette_create(5, df.rgb, 25)
#>   red                 green                blue                 palColors
#> 1 "0.378851676627161" "0.240719238801461"  "0.157756327186512"  "#613D28"
#> 2 "0.131861895991031" "0.0323102245561588" "0.0722739611501933" "#220812"
#> 3 "0.38170335968714"  "0.310329458173915"  "0.631582530290201"  "#614FA1"
#> 4 "0.702260351599259" "0.402291942134016"  "0.467885236256023"  "#B36777"
#> 5 "0.863901473632775" "0.852144075021141"  "0.826650225307372"  "#DCD9D3"